+++ zone/zonedump.h
@@ -104,7 +104,7 @@ struct NPCType
 	uint8	prim_melee_type;
 	uint8	sec_melee_type;
 	uint8	ranged_type;
-	int32	hp_regen;
+	int64	hp_regen;
 	int64	hp_regen_per_second;
 	int32	mana_regen;
 	int32	aggroradius; // added for AI improvement - neotokyo
